Fix function breaking bug in function
The basic idea behind the function is the same, minor changes were made in order to correct it. Changes were made mainly to the route which is explained by the code clearly, and also a conversion was added in order to write back in the same format as it is.